services/requirements.txt
mysql-connector-python>=8.0.0
requests>=2.31.0
cryptography>=41.0.0
python-dotenv>=1.0.0

